zmiana nazw klas na wielkie, wydzielenie klas statków do oddzielnych plików
