+++ loty/src/com/company/State.java
@@ -1,0 +1,11 @@
+package com.company;
+
+import java.awt.*;
+
+public abstract class State {
+    public void lec(double time) {
+
+    }
+
+    public abstract void draw(Graphics2D g2d);
+}
